Revert "Restructure source"
This reverts commit 5fd4e6f3798ca3e43350d7b3ecf9c0e2826e7964.
