+++ src/main/java/bun/ui/BunException.java
@@ -1,4 +1,4 @@
-package bun.exception;
+package bun.ui;
 
 /**
  * BunException is the main entity we'll be using to handle all exceptions specific to Bun.
