+++ src/main/java/bun/ui/Command.java
@@ -1,4 +1,4 @@
-package bun.command;
+package bun.ui;
 
 /**
  * Command is an abstract class for all commands.
