+++ src/main/java/bun/ui/DoAfter.java
@@ -1,4 +1,4 @@
-package bun.task;
+package bun.ui;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
