+++ src/main/java/bun/ui/InvalidIndexException.java
@@ -1,4 +1,4 @@
-package bun.exception;
+package bun.ui;
 
 public class InvalidIndexException extends BunException {
     public InvalidIndexException(int index, int length) {
