+++ src/main/java/bun/ui/Task.java
@@ -1,4 +1,4 @@
-package bun.task;
+package bun.ui;
 
 public class Task {
