+++ src/main/java/bun/ui/TaskList.java
@@ -1,4 +1,4 @@
-package bun.task;
+package bun.ui;
 
 import java.util.ArrayList;
