+++ src/main/java/bun/ui/ToDo.java
@@ -1,4 +1,4 @@
-package bun.task;
+package bun.ui;
 
 public class ToDo extends Task {
     public ToDo(String description) {
